+++ src/connectors/connectoritem.cpp
@@ -475,7 +475,7 @@ ConnectorItem * ConnectorItem::removeConnection(ItemBase * itemBase) {
 			}
 			restoreColor(visited);
 			DebugDialog::debug(QString("remove from:%1 to:%2 count%3")
-			                   .arg((long) this, 0, 16)
+			                   .arg((size_t) this, 0, 16)
 			                   .arg(itemBase->modelPartShared()->title())
 			                   .arg(m_connectedTo.count()) );
 			return removed;
@@ -1944,7 +1944,7 @@ void ConnectorItem::debugInfo(const QString & msg)
 	            .arg(this->attachedToViewLayerPlacement())
 	            .arg(this->attachedTo()->wireFlags())
 	            .arg(this->m_hybrid)
-	            .arg((long) this->bus(), 0, 16)
+	            .arg((size_t) this->bus(), 0, 16)
 	            .arg(this->m_radius)
 	            .arg(this->m_strokeWidth)
 	            .arg(p.x())
